requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
pymysql>=1.1.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
markupsafe>=2.1.0